feat: allow selecting projects
The earlier version of this repository required projects
to be git submodules. This made it hard to re-use this
project for other people, especially when they have no
access to the submodules.

This commit changes the project structure such that
it resides as a neighbor to all projects which shall use
the ai agent rules.